Complete base legacy stories refactor.
