+++ .github/workflows/storybook.yml
@@ -9,12 +9,10 @@ on:
     paths:
       - '.github/workflows/storybook.yml'
       - 'assets/**'
-      - 'stories/**'
       - '!assets/**/__tests__/**/*.js'
       - '!assets/**/test/*.js'
       - '!assets/**/*.test.js'
       - '.storybook/**'
-      - '!.storybook/storybook-data.js'
       - '**.scss'
       - '.nvmrc'
       - '**/package.json'
@@ -28,12 +26,10 @@ on:
     paths:
       - '.github/workflows/storybook.yml'
       - 'assets/**'
-      - 'stories/**'
       - '!assets/**/__tests__/**/*.js'
       - '!assets/**/test/*.js'
       - '!assets/**/*.test.js'
       - '.storybook/**'
-      - '!.storybook/storybook-data.js'
       - '**.scss'
       - '.nvmrc'
       - '**/package.json'
