+++ .storybook/main.js
@@ -32,7 +32,7 @@ function vrtHead() {
 }
 
 module.exports = {
-	stories: [ '../stories/**/*.stories.js', '../assets/js/**/*.stories.js' ],
+	stories: [ '../assets/js/**/*.stories.js' ],
 	addons: [ '@storybook/addon-viewport', '@storybook/addon-postcss' ],
 	previewHead: ( head ) => {
 		if ( process.env.VRT === '1' ) {
